[SYCL][ESIMD] Temporary remove test on fp_extra
